Remove concept of Unmodifiable and Transactional domain logic
Now that we have removed the database and batching logic from
the domain objects it seems unnecessarily complicated to separate
each of EntryLog, ItemStore and RecordIndex into Unmodifiable
and Transactional versions. This now intoduces a single
implementation of each of these.
