+++ src/main/java/uk/gov/register/core/RegisterContext.java
@@ -98,9 +98,9 @@ public class RegisterContext implements
     public Register buildOnDemandRegister() {
         return new PostgresRegister(getRegisterMetadata(),
                 getRegisterFieldsConfiguration(),
-                new UnmodifiableEntryLog(memoizationStore.get(), getOnDemandDataAccessLayer()),
-                new UnmodifiableItemStore(getOnDemandDataAccessLayer()),
-                new UnmodifiableRecordIndex(getOnDemandDataAccessLayer()),
+                new EntryLogImpl(getOnDemandDataAccessLayer(), memoizationStore.get()),
+                new ItemStoreImpl(getOnDemandDataAccessLayer(), new ItemValidator(configManager, registerName)),
+                new RecordIndexImpl(getOnDemandDataAccessLayer()),
                 dbi.onDemand(IndexDAO.class),
                 dbi.onDemand(IndexQueryDAO.class),
                 new DerivationRecordIndex(getOnDemandDataAccessLayer()),
@@ -110,9 +110,9 @@ public class RegisterContext implements
     private Register buildTransactionalRegister(Handle handle, DataAccessLayer dataAccessLayer, TransactionalMemoizationStore memoizationStore) {
         return new PostgresRegister(getRegisterMetadata(),
                 getRegisterFieldsConfiguration(),
-                new TransactionalEntryLog(memoizationStore, dataAccessLayer),
-                new TransactionalItemStore(dataAccessLayer, new ItemValidator(configManager, registerName)),
-                new TransactionalRecordIndex(dataAccessLayer),
+                new EntryLogImpl(dataAccessLayer, memoizationStore),
+                new ItemStoreImpl(dataAccessLayer, new ItemValidator(configManager, registerName)),
+                new RecordIndexImpl(dataAccessLayer),
                 handle.attach(IndexDAO.class),
                 handle.attach(IndexQueryDAO.class),
                 new DerivationRecordIndex(dataAccessLayer),
